removed $datetime info from html-documentation
there is little use in knowing when the documentation was last generated.
(it's way more important to know when the files the docs are based on were
last edited, then when doxygen was last run.)

apart from that: Debian has started the "reproducible builds" initiative,
that attempts to make sure that binaries can be reproduced exactly (given
that you have the same toolchain). this fails if the timestamp of the build
process is part of the build result.

See https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=788757
and https://wiki.debian.org/ReproducibleBuilds