Don't include tracking images
Including an online version of the sourceforge logo in the documentation
means that whenever someone wants to read the offline documentation,
a request is sent to sourceforge.

This is a privacy breach.

Apart from that, I dont even know why you have the sourceforge link:
jacktrip is maintained on github, the mailinglist is run by google groups.