Add files via upload
Finally managed to test on windows, and it did not go well. It seems that either windows sockets or the way that QUdpSocket manages them is not as thread safe as POSIX sockets. It worked when running as a client, but crashed out after about a second as a server. Also, no combination of half closing sockets (the old way of doing it) seemed to work consistently across both IPv4 and IPv6 on all platforms.

So, have moved to sharing a single QUdpSocket object between both the sender and receiver, locking when a read or write is taking place. (Am not currently locking when checking if there are available datagrams, and that doesn't seem to be causing any trouble.)