Revert changes to Hub in external sources
Was just incorporating some of the recent changes to master into my own branch, and it looks like we may have been a little overzealous with the move from master to hub. All of the changes within the JackTrip code look fine, but we probably shouldn't have changed things in the externals directory. Seems like this may have affected the binary file libjack.dll as well, which will likely break things.