QJackTrip - JackTrip with a GUI
I thought I might as well get this out there so that people can at least play around with it and we can iron out the details further down the line. I'd really like to release the GUI part of this under the GPL, so that's what I've done here. (Since it'll be easier for me to re-release the code under a more permissive license later if need be rather than trying to impose a copyleft license after the horse has already bolted.)

Building should have the same requirements as the command line version. There is the option to use CMake, but just running "./build-script" should do the job and output the executable to the build folder. Additionally, in the "os x" folder, there's a script to put together an application bundle. You'll need to make sure the qt bin folder is in your PATH environment variable for it to work.