Check jack client name length
Add check to make sure our jack client name is below the maximum value or below 27 on OS X and truncate it if it's not. (Work around for the jack bug identified in #102.)
This is particularly important for a hub server because the name automatically generated from the IPv6 address can be up to 39 characters long.