Proposed fix to double delete.
Don't delete the ProcessPlugins in the Effects destructor: all of these are passed to the AudioInterface and are deleted there once the JackTrip process ends. I've also moved the allocation for these effects so that it doesn't happen if we're launching a hub server. (So won't cause a memory leak.)