Add time based access to the auth system
Add time checks to logins. Each entry in the auth file is of the format:

username:passwordHash:allowedTimes

A password hash can be generated using "openssl passwd -6"

The allowed times are a comma separated list. Leaving it empty disables acces, and a wildcard entry "*" allows access at any time. Days of the week can be specified in a two letter format followed by a "*" to indicate all day access, or a time range.

So, "We0900-1200,We1400-1530,Th*" would allow access to the server on Wednesday between 9:00am and 12:00pm, and 2:00pm and 3:30pm, and all day on Thursday.