More authentication improvements
Make the location of the credentials store file configurable.

Monitor the credentials file for changes so that users can be added and removed, or access times can be changed without having to restart the hub server.

Minor UI improvements (including changing client and server to P2P client and P2P server for consistency with the command line).

Add a script to generate a certificate, key and initial credentials file for use with hub server authentication.