More adventures in Mach scheduling
Latest attempt to fix the priority of the audio and networking threads. It seems that the scheduler will preempt realtime threads belonging to background apps in an effort to cool the CPU. Attempting to use thread QOS to prevent this, but will need to keep testing and looking into it.