Use Jitter Buffer by Default
If the jitter buffer has previously been disabled, alert the user that it is now the default and enable it.

Add GUI elements to enable broadcast output, and change UDP realtime settings.

In console mode, read the username and password from standard input if they're not provided as arguments. (Disable echo when reading the password.)