Use async networking in hub listener (#150)
* Use async networking in hub listener

Refactored the UdpHubListener to use asynchronous networking. (Should prevent a rogue client from blocking subsequent connections by not sending its port.) This is the last of the Qt socket code to be restructured in this way: the only synchronous networking code is now the bsd/winsock code in UdpDataProtocol.

This change also means that UdpHubListener doesn't require its own runloop, so it is no longer derived from QThread. (SIGINT/SIGTERM are monitored using a timer.)

* Use lambda for callback

* Use QTcpSocket instead of QSslScoket

Prematurely added code from qjacktrip when trying to refactor this.

* Fixed receivedClientInfo definition

This is why I shouldn't try and do things in a hurry.