Restructuring of JackTripWorker and UdpHubListener
JackTripWorker is no longer a QRunnable: using non-blocking, asynchronous networking calls means that we don't need the additional thread. (This was mostly being used for an underutilised eventloop in any case.) UdpHubListener no longer contains any blocking code, so this new structure should solve some of the infinite loop issues that we were seeing when waiting for previous JackTripWorker instances to finish.