Make assemble_app script configurable
By default the script just creates an unsigned app bundle.

Use -i to build an installer.
Supply a certificate using -c to code sign.
Supply a certificate and provide notarization settings using -u -p and -a to send a notarization request to Apple.

Checks whether binary was built with static Qt and doesn't run macdeployqt if it was.