Check for static qt on windows
Checks to see if jacktrip has been built using static qt on windows and only sets path variables if they are not already set.

The script currently assumes the qt mingw tools are installed. If building with msvc, we'll need to copy different compiler dlls into the package, and possibly use dumpbin instead of objdump for the static qt test.