Check if jack is available
If built with weakjack, check if jack is available. If not and the rtaudio backend has been built in, fallback on this and show a warning dialog which the user can choose to hide in the future. If no other backend is available, show a warning that cannot be suppressed.

Also, remove redundant code to change UI based on audio backend choice in the loadSettings function. (This is taken care of when the backendComboBox value is set via the connected signal and slot.)