Fix clazy level 2 warnings
Fixes hundreds of warnings from the clazy level 2 checks. The only check not done is the copyable-polymorphic check. (A lot of the warnings thrown by this are in the faust generated code, and so are probably safe to ignore.) From the remaining checks there are only 7 warnings remaining, and they're in parts of the code that I'm less familiar with. (So I didn't want to break anything.)