RtAudio warning changes
Remove the RtAudio warning for official builds and tweak it for NO_JTVS builds (the binaries maintained by @psiborg112).

Additionally, save if RtAudio has been used as a fallback option rather than selected as a deliberate choice by the user. That way, if JACK is installed later, it can be reinstated as the default backend. (NO_JTVS builds)