Don't connect our UDP socket (#481)
One of the quirks of the pre IPv6 networking code is that we called connect on our UDP socket. The benefit was that we could just call send rather than sendto and it would only accept packets arriving from the specified destination.

I abandoned it for IPv6 because adapters commonly have multiple addresses and there was no guarantee that packets would be arriving back from the same address. I kept it for IPv4 because if it ain't broke... But it is broken. It seems to cause issues with both epoll and kqueue causing them to report that the socket is ready for reading when there's no data on it. Fixes #474.

(Also, clean out some of the very old and very broken 'two sockets on a single port' code.)