Fix check states
Get rid of unnecessary toggle functions and make sure necessary Q_PROPERTY macros exist. Make sure check state is properly restored on load.

Have created NOTIFY signals as well, but not emitting them at the moment in setShowInactive and setShowSelfHosted. (We'll need to add that if we ever need these settings synced elsewhere in the UI.)