Reunify the standard GUI variants
The original GUI was never, ever meant to get as tall as it did, and I always held that moving the auth section to the advanced options tab to appease the VS folk was a major design flaw. (And one that I've been almost entirely insulated from by maintaining my own version of the interface.) Now that they have their own UI, this moves the auth section back where it belongs, with a disclaimer to use the VS interface if that's what you're after.