Adding RegulatorThread to prepare packets from network in parallel to jack callback
Regulator's current implementation runs pullPacket() inside of the jack audio callback.
This works fine most of the time, but I believe it's taking up to 80% of the total
callback interval whenever it has to run predictions. If any other jack clients are
producing a "jack cpu load" of > about 20%, it results in xruns.

This PR introduces a third thread for each client (RegulatorThread), which is used
to continuously run pullPacket() for each clock tick, without impacting the load
on the jack audio callback.

RegulatorThread runs pullPacket() every time the audio callback runs, such that
immediately after a packet is pulled by the callback, it starts preparing the next
one. This allows it run up to the full callback interval without impact the load
on jack.

I also modified the signatures slighly for JackTrip::writeAudioBuffer and
RingBuffer::insertSlotNonBlocking so that specialized versions for Regulator
should not be required.