Replace QVector in meter code
Use a couple of float arrays that only get reallocated if the number of channels changes. (And have a working buffer that is only reallocated if it needs to grow.) This should stop the meter code from churning through memory.

As an added bonus, there's now an option in the classic GUI to execute an external script on connection or disconnection.