Adding universal binary support for OSX builds
Updating qt to the latest open source release (5.15.9) to pull
in latest bug fixes. Supposedly this is also the first release
that adds universal binary support.

Bumping Xcode from 11.7 to 14.0.1 on OSX 12. This uses the latest
minor release of SDK12 (12.3), which qt claims is the only one
tested and known to work with it.

Also bumping minimal target for OSX to 10.13 (High Sierra) which
was released September 25, 2017 (almost 5 years ago). All macs
created after 2009 (over 13 years ago) should support it. This
is necessary because qt seems to have dropped support for all
older releases.

Without these changes, PLC is completely broken in our builds
running on Apple Silicon Macs. A few seconds after connecting,
all output audio drops out and one of the threads goes into an
infinite loop. I was able to identify that this ONLY happens on
Apple Silicon Macs and ONLY when they are running Intel builds.
It doesn't happen anywhere else (Windows, Linux or Intel Macs),
and it doesn't happen running a arm64 build on Apple Silicon
Macs.I suspect that Rosetta has a bug with translation of the
Intel-optimized PLC code that triggers it.