Fixing Play Test Tone button on Linux
It turns out that it is unsafe to include multiple faust-generated
C++ files in the same object. They contain multiple symbols which
use the same name, including APIUI and #ifdef statements. Mixing
these together leads to very unpredictable behaviour, including crashes.

In particular, I discovered the the vsAudioInterface class was including
Tone.h, Meter.h and Volume.h, and each of those was including their own
faust-generated c++ files. This was causing symbol conflicts in the
generated object code. It's actually quite suprising that it only seemed
to have a side-effect on Linux, and only in it "not working." I think
we were just getting lucky.

I've reworked the Tone, Meter and Volume classes so that the faust
generated c++ code is only used in their .cpp files. This ensures that
each of them uses the correct symbols, and the duplicate names do not
cause conflicts.