Optimized performance of virtual studio volume meters
I discovered that the volume meters in virtual studio mode
were using about 3x the CPU versus volume meters in classic
mode. These changes bring them in line with each other.

For more context: I created a stable test case and measured
that JackTrip command line was using about 5% of a CPU core.
JackTrip running in classic mode was using about 10% of a
core, and JackTrip running in virtual studio mode was using
about 20% of a core. This was all with the same traffic,
settings, etc. I narrowed the bulk of the CPU use down to the
volume meters and made various optimizations to avoid the
performance penalty.

There are several changes in here, most of which probably
doesn't make a huge difference. The most impactful change
seems to be using a ListView in the QML but rather two fixed
channels. I suspect this causes the rendering to be a ton
more expensive, but it could also be something else related
to those QML changes.

Updated to remove use of setContextProperty for volume meters