Defer settings and audio initialization outside of QJackTrip constructor
There currently is a strange coupling between the QJackTrip gui and the
Virtual Studio gui, such that they are both running at the same time.
This should probably be addressed, but is for another PR..

This one simple defers the settings initialization for QJackTrip, which
includes RTAudio initialization and scanning, so that it is only done
when the QJackTrip interface is being used.

Previously, QJackTrip would try to run RTAudio device scans even before
the initial window was painted. This likely conflicted with the same
work being performed by the Virtual Studio interface. At best, this
would cause unnecessary delays at startup. At worst, I believe it was
causing crashes on Windows, especially when there are ASIO drivers
present, since that framework seems to be extremely brittle and touchy.