Updating jacktrip.yml to generate new Qt cache entries built using qtbuild.sh and qtbuild.bat
The intention is to not modify any of the existing Qt
cache but only generate new entries for now. The work
to modify the builds to start using them would happen
in subsequent PR that targets the dev branch.

GitHub cache entries are only available to other branches
if they have been merged/generated using the "default"
branch, which for jacktrip is main. So we need to merge
this to main first; otherwise the builds working on
updates in other branches will be extremely slow --
on the order of several hours -- dramatically reducing
our cycle time.