Fixing potential crash on Windows within the AudioSettings dialog.
When an ASIO device is selected and you try to change to another
(ASIO or non-ASIO) device on Windows, it would potentially cause
a crash when using rtaudio to rescan devices while an ASIO
device while already open.

Fixing issue with llvm-12 on ubuntu due to build machine upgrades.

Also removing a redundant call to validatedevices.

Bumping version to 1.10.1 for bugfix release