Fix for failed artifact deployment to google cloud
The 2.0.0-beta1 release artifact failed to automatically deploy
to google cloud when it was released.

https://github.com/jacktriplabs/jacktrip/actions/runs/5790324332/job/15693619792

I believe this is the same problem I ran into working on the Qt
build artifacts. There is a gcloud storage feature that breaks
large file uploads on osx and linux runners.

Copying over the same patch I used to disable this feature,
and will test it with the beta2 release