Removing the screen flicker when you change audio channels or devices
Added a new audio.scanningDevices property to help differentiate
when audio is offline versus device scans. The device scans tend
to be a lot slower versus when we are only restarting audio, so
it probably makes sense to handle them differently.

Added a "Restarting Audio" message next to the device refresh
buttons. This doesn't really apply as well to the change devices
page since it does a reconnect and device scan when you change
anything, and an extra refresh on top when you hit the refresh
button.. But at least this is an improvement and perhaps change
devices can be handled better in another PR..

Also adding re-validation whenever any of the audio settings
are changed, because otherwise you could move from single
input channel to two, or vice versus, and end up with an invalid
mix mode setting.