Updates to handle bad audio interface buffer sizes
Attemping to better handle the case where an audio interface
rejects the buffer size that is configured.

I also updated replaced DEVICE_WARN_LATENCY with
DEVICE_WARN_BUFFER_LATENCY and DEVICE_WARN_ASIO_LATENCY to
differentiate the two high latency scenarios for audio interfaces:
the former is set if you have buffer size > 256 (2x default), and
the latter is set if you have a Windows computer without ASIO
drivers. Both will set the high latency flag, which we plan to
use in the UI to notify someone of the problem.