Don't trigger auth succeeded sequence on token refreshes
This was causing everything to get reset, new device registered,
etc. I'm honestly not sure if this works because it's so hard to
retrigger the condition, but it seems like this should work..

When connecting to a studio, set enabled flag in local device
settings cache. This should help alleviate the problem of it
failing to send heartbeats when the initial connection attempt
gets refused.

Retry establishing heartbeat webconnection when it fails

Disconnect device if the connection to server is lost

Adding a little more debugging output to help troubleshoot
related problems in the future