Embedding windows executable manifest that sets minimum supportted version of 10
Without this, qtwebengine/chromium thinks it is running on Windows
8 and disables a lot of things, including hardware acceleration
for video

Adding AVC license notice for h264 codec. The next release will
include support for h264 video encoding and decoding on Windows
using the openh264 library. This only applies to the video feature
provided by Virtual Studio mode, and is otherwise disabled.