Forcing QtWebEngine to use the OpenGL backend
By default, Chromium will use angle with D3D11. This should
be better, but currently causes really bad video flickering
and other artifacts that make it pretty unusable.

Even the demo Quick Nano browser has these artifacts on
websites like Google Maps, so it's a nasty bug in either
WebEngine or Chromium that they just haven't figured out
yet in 6.5.2.

Bumping JackTrip version to 2.1.0 since things are working
fine in 2.0.2 and it seems unlikely we will need to release
a 2.0.3.