Fix PLC regression causing no audio when connected
This is a regression in 2.1.0-beta that happens maybe 1-10% of the
time at startup and continues until you reset the connection.

mLastSeqNumOut was being used before it was ever set (as -1), and
even worse used as an array index, causing uninitialized memory
read. Suprisingly this wasn't crashing but just returning random
results. If you got unlucky, it put things in a state where it
continuously glitch/underrun, returning no audio.