Updating default behavior for Loss Concealment in Virtual Studio mode
This repurposes the default menu item (2) to be "Loss Concealment (Auto)"
which means it will decide for you whether or not to use a worker thread.
It's based on our latest understanding of what is happening.

Using two different audio interfaces at the same time for input & output
is bad in general for digital audio due to problems with clock skew, and
it seems to be required when using any built-in audio interfaces. On
Macs for example, everything seems to be treated as a different audio
interface: built-in speakers, external headphones, built-in mic, external
mic. The only way to avoid problems is to use an external interface.

At least when using RtAudio, I think this effectively reduces the amount
of time audio callbacks have to complete, because it has to wait for
callbacks from two different devices. If those device are ticking too far
apart, it can leave very little room for JackTrip's callbacks to do much
of anything.

Enter PLC: if you try to run burg predictions inside of the audio
callback, and this interval isn't big enough, it will cause it to skip
audio packets, resulting in static.

PLC's worker thread attempts to alleviate this problem by always having
a packet ready to go, running burg predictions outside of the audio
callback. In rare cases, the window of time available can be too small
for even it to avoid glitches, but it makes it far less likely.

The new "Auto" setting basically just picks one or the other for you,
depending on whether you are using the same interface (no worker) or not
(user worker). You can still explicitly set either, if you want to.