Automatically calculate headroom for PLC
This updates the default behavior for "-q auto" so that it tries
to automatically calculate a headroom value for you. You can
still specify an explicit value, including 0.

The best value depends on the quality of your connection. When
plugged into ethernet, on fiber Internet, 0 works great for me.
Even cross-content. There's no need to add an extra 6 ms of
latency (3 for server and 3 for client) when it doesn't improve
quality.

Switching to wifi is a different story. Stdev goes up from about
0.1 ms to 0.6 ms, and about 3 ms headroom seems like a worthy
tradeoff to minimize glitches. How about other connections? It's
hard to say without being able to do a bunch of testing..

It seems like there should be a way to figure this out
programatically, perhaps by just using the stdev? This branch
takes a stab at one approach that seems to work, but there may
be a much better algorithm out there.