Fix support for sample rates other than 48khz
Fix 2.x virtual studio regression that starting forcing sample
rate of 48000 even if the studio was configured to use a
different one.

Use 44100 instead of 48000 when configuring an audio interface
in virtual studio mode, since this seems to be supported by a
greater range of devices.

Improve error message when an audio interface is disconnected
on Windows computers.