Reuse AudioSettings.qml in ChangeDevice.qml
ChangeDevice had various bugs that had already been fixed in
AudioSettings but not copied over. This resolves the bugs and
makes AudioSettings reusable across all the interfaces, to
prevent more bugs from appearing in the future.

This also has a bunch of various UI tweaks, in particular
with regards to positioning and where the Device Warning
messages are displayed.