Attempting to make the user feedback form easier to dismiss after leaving a virtual studio
We've noticed that people often send "empty feedback" by clicking
on the "Submit" button when no stars are selected and no comments
are written. I believe the "No Thanks" button was being confused
with the "Submit" button when people just wanted to quickly dismiss
it. This should prevent that from being possible by having only
one button, which submits if there is feedback and discards if
there is none.

Hopefully having just one big button will also make it easier for
people to quickly dismiss the dialog.