Work-around for ASIO drivers that have trouble changing sample rates
Some ASIO drivers seem to have issues when you open a stream using
a different sample rate than they are currently using. They will
start the stream but then immediately (and unfortunately,
asynchronously) trigger an error for the sample rate change. It's
unclear if this is a bug in RtAudio, or the driver implementation.
An example driver that does this is "Yamaha Steinberg ASIO USB".

The net result is that audio just stops working, with no feedback
why or what is happening.

This basically just intercepts the errors and triggers a device
refresh when it happens, which works around the issue. It also
ensures that an "Audio Error" warning will be triggered in the
user interface for any audio interface errors that occur after
starting the stream.