Fix for unreliable deeplink handling on Windows
Deep links were failing on Windows about 5% of the time.
Console showed the following messages when it happened:

Never ready to read
Not ready and no bytes available

It seems that QLocalSocket::flush() does bad things, and
docs for it suggest using waitForBytesWritten() instead.
That seems to fix the problem.

I also cleaned up some of the error handling, and made it
support the format jacktrip:join/<studio_id> which is more
comformant with the URI standard.