Defer checks for JACK libraries until opening audio interface
See https://community.jacktrip.org/t/old-jack-crashes-jacktrip-on-windows/722

Old versions of JACK installed on windows causes crashes in any
application that tries to open the client library. This bypasses
the problem for the most part by only ever trying to open the JACK
client library if it is explicitly selected as a backend (which is
not the default, at least for VS mode).

Fix for install directory of opengl32sw.dll on windows

Bumping version to 2.2.4 pending release