Make stronger recommendations about audio devices
As we grow, we've seen people trying to use lots of audio devices
that are old and/or very poor quality. This causes lots of audio
problems that are hard to diagnose and understand. Making stronger
recommendations about devices that we have properly vetted will
lead to more successful outcomes.

Also adding native files for OSX 11, which is required by Qt 6.5
and later.