Ignore devices that look like iPhone microphone
I thought that we may just be able to add a warning to discourage
it, but after implementing and trying it out, it's clearly not
enough. Bad things happen when you select iphone microphone that
go far beyond latency and audio quality.

See https://discussions.apple.com/thread/254349417?sortBy=rank