Getting rid of the 5-second studio refresh timer
It makes for a poor user experience when the list keeps jumping
around, and we already have a button to manually perform a refresh
(in which case it isn't so annoying).

Before it not only would refresh when not desired, it failed to
refresh when you wanted it to (or you had to wait for it, which
made things confusing with status changes). Now it should ensure
a refresh happens immediately, whenever you navigate to the
"browse" window.