Moving Adjust Latency advanced setting into device settings
A new "Audio Quality" slider is now included in the device
settings screen when you are connected to a studio. This
allows you to make changes while connected to a session so
that you can hear the difference on the fly, rather than
having to adjust it blindly.

Also added a display of true buffer latency here so that you
can also see how adjusting this slider changes it. Note that
the setting is used to establish a max latency ceiling, or
tolerance, so your true latency may be lower.

Added support for using studio's setting for queueBuffer. This
is the new default, and the slider to set it manually is only
displayed if this is disabled.