Adding JackTrip Audio Bridge VST3 Plugin
Currently this only builds a VST plugin that does nothing, based
on Steinberg's Hello World example. It should work across platforms
and load OK in all hosts that support VST3.

This plugin is intentionally as simple as possible. It supports
two input channels and two output channels, and has no user
interface. The audio signal it receives will be mixed into your
JackTrip input audio signal, and it will output the audio signal
received back from a remote JackTrip server. There is likely a
lot more that can be added over time, but this is the MVP.

Support for building the VST3 plugin is enabled by providing
the following two parameters to meson setup:

* vst-sdkdir: directory where the VST3 SDK is installed (headers)
* vst-libdir: directory where compiled VST3 libraries are located